Gene set analysis using pre-computed test statistic

Perform gene set analysis on the result of a pre-computed test statistic. Test whether statistics in a gene set are larger/smaller than statistics not in the set.

Usage

zenithPR_gsa(
  statistics,
  ids,
  geneSets,
  use.ranks = FALSE,
  n_genes_min = 10,
  progressbar = TRUE,
  inter.gene.cor = 0.01
)

Arguments

statistics

pre-computed test statistics

ids

name of gene for each entry in statistics

geneSets

GeneSetCollection

use.ranks

do a rank-based test TRUE or a parametric test FALSE? default: FALSE

n_genes_min

minumum number of genes in a geneset

progressbar

if TRUE, show progress bar

inter.gene.cor

correlation of test statistics with in gene set

Value

NGenes: number of genes in this set

Correlation: mean correlation between expression of genes in this set

delta: difference in mean t-statistic for genes in this set compared to genes not in this set

se: standard error of delta

p.less: p-value for hypothesis test of H0: delta < 0

p.greater: p-value for hypothesis test of H0: delta > 0

PValue: p-value for hypothesis test H0: delta != 0

Direction: direction of effect based on sign(delta)

FDR: false discovery rate based on Benjamini-Hochberg method in p.adjust

Details

This is the same as zenith_gsa(), but uses pre-computed test statistics. Note that zenithPR_gsa() may give slightly different results for small samples sizes, if zenithPR_gsa() is fed t-statistics instead of z-statistics.

See also